requirements.txt
fastapi
uvicorn
pillow
numpy
joblib
scikit-learn
